Reorganize files according to new structure
